Add MarvelApiService DI
