+++ app/src/main/java/com/example/internapp/BaseApplication.kt
@@ -1,0 +1,11 @@
+package com.example.internapp
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class BaseApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+    }
+}
