+++ app/src/main/java/com/example/internapp/repository/MarvelModule.kt
@@ -1,0 +1,38 @@
+package com.example.internapp.repository
+
+import dagger.Provides
+import dagger.Module
+import dagger.hilt.InstallIn
+import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
+import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import dagger.hilt.android.components.ActivityRetainedComponent
+import okhttp3.OkHttpClient
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
+import javax.inject.Singleton
+
+@Module
+@InstallIn(ActivityRetainedComponent::class)
+object MarvelModule {
+    private const val BASE_URL = "https://gateway.marvel.com/"
+    private val builder = OkHttpClient.Builder()
+        .connectTimeout(20, java.util.concurrent.TimeUnit.SECONDS)
+        .writeTimeout(30, java.util.concurrent.TimeUnit.SECONDS)
+        .readTimeout(30, java.util.concurrent.TimeUnit.SECONDS)
+    private val client = builder.build()
+
+    private val moshi = Moshi.Builder()
+        .add(KotlinJsonAdapterFactory())
+        .build()
+
+    private val retrofit = Retrofit.Builder()
+        .baseUrl(BASE_URL)
+        .addConverterFactory(MoshiConverterFactory.create(moshi))
+        .addCallAdapterFactory(CoroutineCallAdapterFactory())
+        .client(client)
+        .build()
+
+    @Provides
+    fun provideMarvelApiService(): MarvelApiService = retrofit.create(MarvelApiService::class.java)
+}
