+++ app/src/main/java/com/example/internapp/ui/HomeFragment.kt
@@ -9,7 +9,9 @@ import com.example.internapp.MainViewModel
 import com.example.internapp.R
 import com.example.internapp.databinding.HomeFragmentBinding
 import com.example.internapp.repository.UIState
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class HomeFragment : Fragment() {
     private val viewModel: MainViewModel by activityViewModels()
     private lateinit var binding: HomeFragmentBinding
